+++ lib/core/mergeConfig.js
@@ -33,11 +33,11 @@ module.exports = function mergeConfig(config1, config2) {
 
   utils.forEach(mergeDeepPropertiesKeys, function mergeDeepProperties(prop) {
     if (utils.isObject(config2[prop])) {
-      config[prop] = utils.deepMerge(config1[prop], config2[prop]);
+      config[prop] = utils.merge(config1[prop], config2[prop]);
     } else if (typeof config2[prop] !== 'undefined') {
       config[prop] = config2[prop];
     } else if (utils.isObject(config1[prop])) {
-      config[prop] = utils.deepMerge(config1[prop]);
+      config[prop] = utils.merge(config1[prop]);
     } else if (typeof config1[prop] !== 'undefined') {
       config[prop] = config1[prop];
     }
